[Refact] #5 ReissueToken 방식 Json으로 변경

+++ src/main/java/com/WhoIsRoom/WhoIs_Server/domain/auth/service/JwtService.java
@@ -1,18 +1,24 @@
 package com.WhoIsRoom.WhoIs_Server.domain.auth.service;
 
 import com.WhoIsRoom.WhoIs_Server.domain.auth.dto.request.RefreshTokenRequest;
+import com.WhoIsRoom.WhoIs_Server.domain.auth.dto.response.LoginResponse;
+import com.WhoIsRoom.WhoIs_Server.domain.auth.dto.response.ReissueResponse;
 import com.WhoIsRoom.WhoIs_Server.domain.auth.exception.CustomAuthenticationException;
 import com.WhoIsRoom.WhoIs_Server.domain.auth.exception.CustomJwtException;
 import com.WhoIsRoom.WhoIs_Server.domain.auth.util.JwtUtil;
 import com.WhoIsRoom.WhoIs_Server.global.common.redis.RedisService;
+import com.WhoIsRoom.WhoIs_Server.global.common.response.BaseResponse;
 import com.WhoIsRoom.WhoIs_Server.global.common.response.ErrorCode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
 import java.time.Duration;
 
 @Slf4j
@@ -38,24 +44,30 @@ public class JwtService {
 
     private final RedisService redisService;
     private final JwtUtil jwtUtil;
+    private final ObjectMapper objectMapper;
 
     public void logout(HttpServletRequest request, RefreshTokenRequest tokenRequest) {
         String accessToken = jwtUtil.extractAccessToken(request)
                 .orElseThrow(() -> new CustomAuthenticationException(ErrorCode.SECURITY_INVALID_ACCESS_TOKEN));
+
         String refreshToken = tokenRequest.getRefreshToken();
+        jwtUtil.validateToken(refreshToken);
+        if (!"refresh".equals(jwtUtil.getTokenType(refreshToken))) {
+            throw new CustomJwtException(ErrorCode.INVALID_TOKEN_TYPE);
+        }
 
         deleteRefreshToken(refreshToken);
         //access token blacklist 처리 -> 로그아웃한 사용자가 요청 시 access token이 redis에 존재하면 jwtAuthenticationFilter에서 인증처리 거부
         invalidAccessToken(accessToken);
     }
 
-    public void reissueTokens(HttpServletRequest request, HttpServletResponse response, RefreshTokenRequest tokenRequest) {
+    public ReissueResponse reissueTokens(RefreshTokenRequest tokenRequest) {
         String refreshToken = tokenRequest.getRefreshToken();
         jwtUtil.validateToken(refreshToken);
         if (!"refresh".equals(jwtUtil.getTokenType(refreshToken))) {
             throw new CustomJwtException(ErrorCode.INVALID_TOKEN_TYPE);
         }
-        reissueAndSendTokens(response, refreshToken);
+        return reissueAndSendTokens(refreshToken);
     }
 
     public void checkLogout(String accessToken) {
@@ -81,7 +93,7 @@ public class JwtService {
                 Duration.ofMillis(ACCESS_TOKEN_EXPIRED_IN));
     }
 
-    private void reissueAndSendTokens(HttpServletResponse response, String refreshToken) {
+    private ReissueResponse reissueAndSendTokens(String refreshToken) {
 
         // 새로운 Refresh Token 발급
         String reissuedAccessToken = jwtUtil.createAccessToken(jwtUtil.getUserId(refreshToken), jwtUtil.getProviderId(refreshToken), jwtUtil.getRole(refreshToken), jwtUtil.getName(refreshToken));
@@ -93,12 +105,9 @@ public class JwtService {
         // 기존 Refresh Token 폐기 (DB나 Redis에서 삭제)
         deleteRefreshToken(refreshToken);
 
-        sendTokens(response, reissuedAccessToken, reissuedRefreshToken);
-    }
-
-    public void sendTokens(HttpServletResponse response, String accessToken,
-                            String refreshToken) {
-        response.setHeader(ACCESS_HEADER, BEARER_PREFIX + accessToken);
-        response.setHeader(REFRESH_HEADER, BEARER_PREFIX + refreshToken);
+        return ReissueResponse.builder()
+                .accessToken(reissuedAccessToken)
+                .refreshToken(reissuedRefreshToken)
+                .build();
     }
 }
